+++ packages/ui-components/src/__tests__/ListViewOrderbookFilters.test.ts
@@ -105,27 +105,4 @@ describe('ListViewOrderbookFilters', () => {
 		expect(screen.queryByTestId('order-hash-input')).not.toBeInTheDocument();
 		expect(screen.queryByTestId('order-status-dropdown')).not.toBeInTheDocument();
 	});
-
-	test('refetches data when refresh button is clicked', async () => {
-		const mockRefetch = vi.fn();
-		const mockQuery = writable({
-			status: 'success',
-			fetchStatus: 'idle',
-			refetch: mockRefetch
-			// ... other required properties
-		});
-
-		render(ListViewOrderbookFilters, {
-			...defaultProps,
-			query: mockQuery as unknown as CreateInfiniteQueryResult<
-				InfiniteData<unknown[], unknown>,
-				Error
-			>
-		});
-
-		const refreshButton = screen.getByTestId('refresh-button');
-		await userEvent.click(refreshButton);
-
-		expect(mockRefetch).toHaveBeenCalled();
-	});
 });
